# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RedSecTools API",
    "version": "2.0.0",
    "description": "Repository-only OpenAPI documentation for RedSecTools. This file is not served by the live application. It documents the current HTTP API, including auth requirements for main app, guest-link, extension, and admin routes. RedSecThreat now includes user-scoped keyword, tag, and alert ownership on top of shared feed ingestion, plus admin-managed feed, template, and notification policy routes. Survey and Wiki include their active create/manage/respond/results/documentation route groups alongside the existing Paste, Share, Vault, Chat, Calendar, Homepage, Extension, and Admin surfaces."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_session",
        "description": "Signed user session cookie for the main web app."
      },
      "GuestSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_guest",
        "description": "Signed guest cookie created from /guest/{token}. Tool-scoped for paste or share."
      },
      "ExtensionBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque token",
        "description": "Bearer token returned by /api/ext/auth/login or /api/ext/auth/login/mfa."
      },
      "AdminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_admin",
        "description": "Signed admin session cookie for /admin/api routes."
      }
    },
    "schemas": {
...
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        },
        "required": ["success"]
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or expired",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/threat/bootstrap": {
      "get": {
        "tags": ["Threat"],
        "summary": "Bootstrap threat intel dashboard",
        "description": "Returns the current user's threat dashboard bootstrap payload, including user-scoped keyword and alert stats, recent user-owned alerts, shared feed health, and personal notification preferences. Requires threat.view permission.",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Dashboard bootstrap data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stats": { "type": "object", "description": "Alert counts by criticality and total feed/keyword counts" },
                    "recentAlerts": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatAlert" } },
                    "feedHealth": { "type": "array", "items": { "type": "object" } },
                    "settings": {
                      "type": "object",
                      "properties": {
                        "autoFetchEnabled": { "type": "boolean" },
                        "fetchIntervalSeconds": { "type": "integer" }
                      }
                    },
                    "userNotifications": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "403": { "description": "Threat intel access denied (missing threat.view permission)" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/feeds": {
      "get": {
        "tags": ["Threat"],
        "summary": "List threat feeds",
        "description": "Returns all configured feed sources. Query param ?enabled=true filters to enabled only. Requires threat.view.",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          { "name": "enabled", "in": "query", "required": false, "schema": { "type": "string", "enum": ["true", "1"] }, "description": "Filter to enabled feeds only" }
        ],
        "responses": {
          "200": { "description": "Feed list", "content": { "application/json": { "schema": { "type": "object", "properties": { "feeds": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatFeed" } } } } } } },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Threat"],
        "summary": "Create a threat feed",
        "description": "Public user sessions cannot create feeds here. Feed source management lives in the admin panel under /admin/api/threat/feeds.",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreatFeedRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Feed created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "feed": { "$ref": "#/components/schemas/ThreatFeed" } } } } } },
          "400": { "description": "Validation error (missing name/url, invalid feed type)" },
          "403": { "description": "Feed source changes are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/feeds/{id}": {
      "get": {
        "tags": ["Threat"],
        "summary": "Get a single threat feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Feed details", "content": { "application/json": { "schema": { "type": "object", "properties": { "feed": { "$ref": "#/components/schemas/ThreatFeed" } } } } } },
          "404": { "description": "Feed not found" },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["Threat"],
        "summary": "Update a threat feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThreatFeedRequest" } } }
        },
        "responses": {
          "200": { "description": "Feed updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "feed": { "$ref": "#/components/schemas/ThreatFeed" } } } } } },
          "400": { "description": "Validation error" },
          "404": { "description": "Feed not found" },
          "403": { "description": "Management denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Threat"],
        "summary": "Delete a threat feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Feed deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Success" } } } },
          "400": { "description": "Invalid feed ID" },
          "403": { "description": "Management denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/feeds/{id}/check": {
      "post": {
        "tags": ["Threat"],
        "summary": "Manually trigger a feed check",
        "description": "Public user sessions cannot trigger feed checks here. Use the admin panel feed refresh controls instead.",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Feed check completed", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "alerts": { "type": "array", "items": { "type": "object" } } } } } } },
          "400": { "description": "Invalid feed ID" },
          "500": { "description": "Feed check failed" },
          "403": { "description": "Feed source changes are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/feeds/{id}/keywords": {
      "post": {
        "tags": ["Threat"],
        "summary": "Set keyword associations for a feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "keywordIds": { "type": "array", "items": { "type": "string" } } }, "required": ["keywordIds"] } } }
        },
        "responses": {
          "200": { "description": "Keywords updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "keywords": { "type": "array", "items": { "type": "object" } } } } } } },
          "404": { "description": "Feed not found" },
          "403": { "description": "Feed source changes are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/keywords": {
      "get": {
        "tags": ["Threat"],
        "summary": "List threat keywords",
        "description": "Returns the current user's effective keyword list, including shared default keywords plus personal keywords. Shared defaults can be disabled per user without being deleted. Query param ?enabled=true filters to enabled only. Requires threat.view.",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          { "name": "enabled", "in": "query", "required": false, "schema": { "type": "string", "enum": ["true", "1"] }, "description": "Filter to enabled keywords only" }
        ],
        "responses": {
          "200": { "description": "Keyword list", "content": { "application/json": { "schema": { "type": "object", "properties": { "keywords": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatKeyword" } } } } } } },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Threat"],
        "summary": "Create a threat keyword",
        "description": "Creates a personal keyword rule for the current user with optional regex support and criticality level. Requires threat.view.",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ThreatKeywordRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Keyword created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "keyword": { "$ref": "#/components/schemas/ThreatKeyword" } } } } } },
          "400": { "description": "Validation error (empty keyword, too long)" },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/threat/tags": {
      "get": {
        "tags": ["Threat"],
        "summary": "List all threat tags",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": { "description": "Tag list", "content": { "application/json": { "schema": { "type": "object", "properties": { "tags": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatTag" } } } } } } },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Threat"],
        "summary": "Create a threat tag",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "color": { "type": "string", "description": "Hex colour code", "default": "#E53935" },
                  "description": { "type": ["string", "null"] }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tag created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "tag": { "$ref": "#/components/schemas/ThreatTag" } } } } } },
          "400": { "description": "Validation error (empty name, too long)" },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}